move a bit of functionality to separate files and add tests
