+++ components/function-controller/internal/controllers/serverless/function_controller.go
@@ -10,145 +10,16 @@ import (
 	"strings"
 
 	"github.com/go-logr/logr"
+	serverlessv1alpha1 "github.com/kyma-project/kyma/components/function-controller/pkg/apis/serverless/v1alpha1"
 	batchv1 "k8s.io/api/batch/v1"
 	corev1 "k8s.io/api/core/v1"
-	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	apilabels "k8s.io/apimachinery/pkg/labels"
-	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/selection"
-	"k8s.io/client-go/tools/record"
 	"knative.dev/pkg/apis"
 	servingv1 "knative.dev/serving/pkg/apis/serving/v1"
 	ctrl "sigs.k8s.io/controller-runtime"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-
-	"github.com/kyma-project/kyma/components/function-controller/internal/resource"
-	serverlessv1alpha1 "github.com/kyma-project/kyma/components/function-controller/pkg/apis/serverless/v1alpha1"
 )
-
-const (
-	functionNameLabel      = "serverless.kyma-project.io/function-name"
-	functionManagedByLabel = "serverless.kyma-project.io/managed-by"
-	functionUUIDLabel      = "serverless.kyma-project.io/uuid"
-
-	serviceBindingUsagesAnnotation = "servicebindingusages.servicecatalog.kyma-project.io/tracing-information"
-
-	cfgGenerationLabel = "serving.knative.dev/configurationGeneration"
-
-	configMapFunction = "handler.js"
-	configMapHandler  = "handler.main"
-	configMapDeps     = "package.json"
-)
-
-var (
-	envVarsForRevision = []corev1.EnvVar{
-		{Name: "FUNC_HANDLER", Value: "main"},
-		{Name: "MOD_NAME", Value: "handler"},
-		{Name: "FUNC_TIMEOUT", Value: "180"},
-		{Name: "FUNC_RUNTIME", Value: "nodejs12"},
-		// {Name: "FUNC_MEMORY_LIMIT", Value: "128Mi"},
-		{Name: "FUNC_PORT", Value: "8080"},
-		{Name: "NODE_PATH", Value: "$(KUBELESS_INSTALL_VOLUME)/node_modules"},
-	}
-)
-
-type FunctionReconciler struct {
-	client.Client
-	Log logr.Logger
-
-	resourceClient resource.Resource
-	recorder       record.EventRecorder
-	config         FunctionConfig
-	scheme         *runtime.Scheme
-}
-
-func NewFunction(client client.Client, log logr.Logger, config FunctionConfig, scheme *runtime.Scheme, recorder record.EventRecorder) *FunctionReconciler {
-	resourceClient := resource.New(client, scheme)
-
-	return &FunctionReconciler{
-		Client:         client,
-		Log:            log,
-		config:         config,
-		resourceClient: resourceClient,
-		scheme:         scheme,
-		recorder:       recorder,
-	}
-}
-
-func (r *FunctionReconciler) SetupWithManager(mgr ctrl.Manager) error {
-	return ctrl.NewControllerManagedBy(mgr).
-		For(&serverlessv1alpha1.Function{}).
-		Owns(&corev1.ConfigMap{}).
-		Owns(&batchv1.Job{}).
-		Owns(&servingv1.Service{}).
-		Owns(&servingv1.Revision{}).
-		Complete(r)
-}
-
-// Reconcile reads that state of the cluster for a Function object and makes changes based on the state read and what is in the Function.Spec
-// +kubebuilder:rbac:groups="",resources=configmaps,verbs=get;list;watch;create;update;patch;delete
-// +kubebuilder:rbac:groups="serverless.kyma-project.io",resources=functions,verbs=get;list;watch;create;update;patch;delete
-// +kubebuilder:rbac:groups="serverless.kyma-project.io",resources=functions/status,verbs=get;update;patch
-// +kubebuilder:rbac:groups="serving.knative.dev",resources=services,verbs=get;list;watch;create;update;patch;delete
-// +kubebuilder:rbac:groups="serving.knative.dev",resources=services/status,verbs=get
-// +kubebuilder:rbac:groups="batch",resources=jobs,verbs=get;list;watch;create;update;patch;delete;deletecollection
-// +kubebuilder:rbac:groups="batch",resources=jobs/status,verbs=get
-// +kubebuilder:rbac:groups="",resources=events,verbs=create;patch
-func (r *FunctionReconciler) Reconcile(request ctrl.Request) (ctrl.Result, error) {
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	instance := &serverlessv1alpha1.Function{}
-	err := r.Get(ctx, request.NamespacedName, instance)
-	if err != nil {
-		if apierrors.IsNotFound(err) {
-			return ctrl.Result{}, nil
-		}
-		return ctrl.Result{}, err
-	}
-
-	log := r.Log.WithValues("kind", instance.GetObjectKind().GroupVersionKind().Kind, "name", instance.GetName(), "namespace", instance.GetNamespace(), "version", instance.GetGeneration())
-
-	log.Info("Listing ConfigMaps")
-	var configMaps corev1.ConfigMapList
-	if err := r.resourceClient.ListByLabel(ctx, instance.GetNamespace(), r.functionLabels(instance), &configMaps); err != nil {
-		log.Error(err, "Cannot list ConfigMaps")
-		return ctrl.Result{}, err
-	}
-	log.Info("Listing Jobs")
-	var jobs batchv1.JobList
-	if err := r.resourceClient.ListByLabel(ctx, instance.GetNamespace(), r.functionLabels(instance), &jobs); err != nil {
-		log.Error(err, "Cannot list Jobs")
-		return ctrl.Result{}, err
-	}
-	log.Info("Gathering Service")
-	service := &servingv1.Service{}
-	if err := r.Client.Get(ctx, request.NamespacedName, service); err != nil {
-		if apierrors.IsNotFound(err) {
-			service = nil
-		} else {
-			log.Error(err, "Cannot get Service %s", instance.GetName())
-			return ctrl.Result{}, err
-		}
-	}
-
-	log.Info("Listing Revisions")
-	var revisions servingv1.RevisionList
-	if err := r.resourceClient.ListByLabel(ctx, instance.GetNamespace(), r.functionLabels(instance), &revisions); err != nil {
-		log.Error(err, "Cannot list Revisions")
-		return ctrl.Result{}, err
-	}
-
-	switch {
-	case r.isOnConfigMapChange(instance, configMaps.Items, service):
-		return r.onConfigMapChange(ctx, log, instance, configMaps.Items)
-	case r.isOnJobChange(instance, jobs.Items, service):
-		return r.onJobChange(ctx, log, instance, configMaps.Items[0].GetName(), jobs.Items)
-	default:
-		return r.onServiceChange(ctx, log, instance, service, revisions.Items)
-	}
-}
 
 func (r *FunctionReconciler) isOnConfigMapChange(instance *serverlessv1alpha1.Function, configMaps []corev1.ConfigMap, service *servingv1.Service) bool {
 	image := r.buildExternalImageAddress(instance)
@@ -384,7 +255,7 @@ func (r *FunctionReconciler) updateService(ctx context.Context, log logr.Logger,
 func (r *FunctionReconciler) deleteRevisions(ctx context.Context, log logr.Logger, instance *serverlessv1alpha1.Function, service *servingv1.Service, revisions []servingv1.Revision) (ctrl.Result, error) {
 	log.Info("Deleting all old revisions")
 
-	selector, err := r.getOldRevisionLabel(instance, revisions)
+	selector, err := r.getOldRevisionSelector(instance, revisions)
 	if err != nil {
 		log.Error(err, "Cannot create proper selector for old revisions")
 		return ctrl.Result{}, err
@@ -620,171 +491,7 @@ func (r *FunctionReconciler) sanitizeDependencies(dependencies string) string {
 	return result
 }
 
-func (r *FunctionReconciler) buildJob(instance *serverlessv1alpha1.Function, configMapName string) batchv1.Job {
-	imageName := r.buildInternalImageAddress(instance)
-	one := int32(1)
-	zero := int32(0)
-
-	job := batchv1.Job{
-		ObjectMeta: metav1.ObjectMeta{
-			GenerateName: fmt.Sprintf("%s-build-", instance.GetName()),
-			Namespace:    instance.GetNamespace(),
-			Labels:       r.functionLabels(instance),
-		},
-		Spec: batchv1.JobSpec{
-			Parallelism:           &one,
-			Completions:           &one,
-			ActiveDeadlineSeconds: nil,
-			BackoffLimit:          &zero,
-			Template: corev1.PodTemplateSpec{
-				ObjectMeta: metav1.ObjectMeta{
-					Labels: r.functionLabels(instance),
-					Annotations: map[string]string{
-						"sidecar.istio.io/inject": "false",
-					},
-				},
-				Spec: corev1.PodSpec{
-					Volumes: []corev1.Volume{
-						{
-							Name: "sources",
-							VolumeSource: corev1.VolumeSource{
-								ConfigMap: &corev1.ConfigMapVolumeSource{
-									LocalObjectReference: corev1.LocalObjectReference{Name: configMapName},
-								},
-							},
-						},
-						{
-							Name: "runtime",
-							VolumeSource: corev1.VolumeSource{
-								ConfigMap: &corev1.ConfigMapVolumeSource{
-									LocalObjectReference: corev1.LocalObjectReference{Name: r.config.Build.RuntimeConfigMapName},
-								},
-							},
-						},
-						{
-							Name: "credentials",
-							VolumeSource: corev1.VolumeSource{
-								Secret: &corev1.SecretVolumeSource{SecretName: r.config.ImagePullSecretName},
-							},
-						},
-						{
-							Name:         "tekton-home",
-							VolumeSource: corev1.VolumeSource{EmptyDir: &corev1.EmptyDirVolumeSource{}},
-						},
-						{
-							Name:         "tekton-workspace",
-							VolumeSource: corev1.VolumeSource{EmptyDir: &corev1.EmptyDirVolumeSource{}},
-						},
-					},
-					InitContainers: []corev1.Container{
-						{
-							Name:    "credential-initializer",
-							Image:   r.config.Build.CredsInitImage,
-							Command: []string{"/ko-app/creds-init"},
-							Args:    []string{fmt.Sprintf("-basic-docker=credentials=http://%s", imageName)},
-							Env: []corev1.EnvVar{
-								{Name: "HOME", Value: "/tekton/home"},
-							},
-							VolumeMounts: []corev1.VolumeMount{
-								{Name: "tekton-home", ReadOnly: false, MountPath: "/tekton/home"},
-								{Name: "credentials", ReadOnly: false, MountPath: "/tekton/creds-secrets/credentials"},
-							},
-							ImagePullPolicy: corev1.PullIfNotPresent,
-						},
-					},
-					Containers: []corev1.Container{
-						{
-							Name:  "executor",
-							Image: r.config.Build.ExecutorImage,
-							Args:  []string{fmt.Sprintf("--destination=%s", imageName), "--insecure", "--skip-tls-verify"},
-							Resources: corev1.ResourceRequirements{
-								Limits: corev1.ResourceList{
-									corev1.ResourceMemory: r.config.Build.LimitsMemoryValue,
-									corev1.ResourceCPU:    r.config.Build.LimitsCPUValue,
-								},
-								Requests: corev1.ResourceList{
-									corev1.ResourceMemory: r.config.Build.RequestsMemoryValue,
-									corev1.ResourceCPU:    r.config.Build.RequestsCPUValue,
-								},
-							},
-							VolumeMounts: []corev1.VolumeMount{
-								{Name: "sources", ReadOnly: true, MountPath: "/src"},
-								{Name: "runtime", ReadOnly: true, MountPath: "/workspace"},
-								{Name: "tekton-home", ReadOnly: false, MountPath: "/tekton/home"},
-							},
-							ImagePullPolicy: corev1.PullIfNotPresent,
-							Env: []corev1.EnvVar{
-								{Name: "DOCKER_CONFIG", Value: "/tekton/home/.docker/"},
-							},
-						},
-					},
-					RestartPolicy:      corev1.RestartPolicyNever,
-					ServiceAccountName: r.config.ImagePullAccountName,
-				},
-			},
-		},
-	}
-
-	return job
-}
-
-func (r *FunctionReconciler) buildService(log logr.Logger, instance *serverlessv1alpha1.Function, oldService *servingv1.Service) servingv1.Service {
-	imageName := r.buildExternalImageAddress(instance)
-	annotations := map[string]string{
-		"autoscaling.knative.dev/minScale": "1",
-		"autoscaling.knative.dev/maxScale": "1",
-	}
-	if instance.Spec.MinReplicas != nil {
-		annotations["autoscaling.knative.dev/minScale"] = fmt.Sprintf("%d", *instance.Spec.MinReplicas)
-	}
-	if instance.Spec.MaxReplicas != nil {
-		annotations["autoscaling.knative.dev/maxScale"] = fmt.Sprintf("%d", *instance.Spec.MaxReplicas)
-	}
-	serviceLabels := r.functionLabels(instance)
-	serviceLabels["serving.knative.dev/visibility"] = "cluster-local"
-
-	bindingAnnotation := ""
-	if oldService != nil {
-		bindingAnnotation = oldService.GetAnnotations()[serviceBindingUsagesAnnotation]
-	}
-	podLabels := r.servingPodLabels(log, instance, bindingAnnotation)
-
-	service := servingv1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      instance.GetName(),
-			Namespace: instance.GetNamespace(),
-			Labels:    serviceLabels,
-		},
-		Spec: servingv1.ServiceSpec{
-			ConfigurationSpec: servingv1.ConfigurationSpec{
-				Template: servingv1.RevisionTemplateSpec{
-					ObjectMeta: metav1.ObjectMeta{
-						Annotations: annotations,
-						Labels:      podLabels,
-					},
-					Spec: servingv1.RevisionSpec{
-						PodSpec: corev1.PodSpec{
-							Containers: []corev1.Container{
-								{
-									Name:            "lambda",
-									Image:           imageName,
-									Env:             append(instance.Spec.Env, envVarsForRevision...),
-									Resources:       instance.Spec.Resources,
-									ImagePullPolicy: corev1.PullIfNotPresent,
-								},
-							},
-							ServiceAccountName: r.config.ImagePullAccountName,
-						},
-					},
-				},
-			},
-		},
-	}
-
-	return service
-}
-
-func getMaxGeneration(revisions []servingv1.Revision) (int, error) {
+func getNewestGeneration(revisions []servingv1.Revision) (int, error) {
 	maxGeneration := -1
 	for _, revision := range revisions {
 		generationString, ok := revision.Labels[cfgGenerationLabel]
@@ -804,8 +511,8 @@ func getMaxGeneration(revisions []servingv1.Revision) (int, error) {
 	return maxGeneration, nil
 }
 
-func (r *FunctionReconciler) getOldRevisionLabel(instance *serverlessv1alpha1.Function, revisions []servingv1.Revision) (apilabels.Selector, error) {
-	maxGen, err := getMaxGeneration(revisions)
+func (r *FunctionReconciler) getOldRevisionSelector(instance *serverlessv1alpha1.Function, revisions []servingv1.Revision) (apilabels.Selector, error) {
+	maxGen, err := getNewestGeneration(revisions)
 	if err != nil {
 		return nil, err
 	}
